+++ Apps/Contoso.Android.Puppet/MainActivity.cs
@@ -18,7 +18,7 @@ namespace Contoso.Android.Puppet
 {
     using AlertDialog = global::Android.App.AlertDialog;
 
-    [Activity(Exported = true, Label = "SXPuppet", Icon = "@drawable/icon", Theme = "@style/PuppetTheme", MainLauncher = true, ConfigurationChanges = ConfigChanges.ScreenSize | ConfigChanges.Orientation)]
+    [Activity(Label = "SXPuppet", Icon = "@drawable/icon", Theme = "@style/PuppetTheme", MainLauncher = true, ConfigurationChanges = ConfigChanges.ScreenSize | ConfigChanges.Orientation)]
     [System.Obsolete]
     public class MainActivity : AppCompatActivity, BottomNavigationView.IOnNavigationItemSelectedListener
     {
