+++ Apps/Contoso.Forms.Demo/Contoso.Forms.Demo.MacOS/AppDelegate.cs
@@ -3,22 +3,22 @@
 
 using System;
 using AppKit;
+using Contoso.Forms.Demo;
 using Foundation;
+using Xamarin.Forms;
 using Xamarin.Forms.Platform.MacOS;
 
 namespace Contoso.Forms.Demo.MacOS
 {
     [Register("AppDelegate")]
-    public class AppDelegate : FormsApplicationDelegate, IClearCrashClick, IAppConfiguration
+    public class AppDelegate : FormsApplicationDelegate, IClearCrashClick
     {
         private const string CrashesUserConfirmationStorageKey = "MSAppCenterCrashesUserConfirmation";
         private const string ApplicationCrashOnExceptionsKey = "NSApplicationCrashOnExceptions";
-        private static string AppCenterSecret;
 
         NSPanel window;
         public AppDelegate()
         {
-            AppCenterSecret = Environment.GetEnvironmentVariable("XAMARIN_FORMS_MACOS_PROD");
             var style = NSWindowStyle.Closable | NSWindowStyle.Resizable | NSWindowStyle.Titled;
             var rect = new CoreGraphics.CGRect(200, 1000, 1024, 768);
             window = new NSPanel(rect, style, NSBackingStore.Buffered, false);
@@ -44,16 +44,6 @@ namespace Contoso.Forms.Demo.MacOS
         public void ClearCrashButton()
         {
             NSUserDefaults.StandardUserDefaults.RemoveObject(CrashesUserConfirmationStorageKey);
-        }
-
-        public string GetAppSecret()
-        {
-            return AppCenterSecret;
-        }
-
-        public string GetTargetToken()
-        {
-            return "";
         }
     }
 }
