+++ Apps/Contoso.Forms.Demo/Contoso.Forms.Demo.iOS/Contoso.Forms.Demo.iOS.csproj
@@ -80,6 +80,12 @@
     <MtouchTlsProvider>Default</MtouchTlsProvider>
     <CodesignProvision>appcenter xamarin forms ios demo df Dist</CodesignProvision>
   </PropertyGroup>
+  <PropertyGroup Condition=" '$(RunConfiguration)' == 'Default' ">
+    <EnvironmentVariables>
+      <Variable name="XAMARIN_FORMS_IOS_PROD" value="app-secret" />
+      <Variable name="XAMARIN_FORMS_IOS_TARGET_TOKEN_PROD" value="target-token" />
+    </EnvironmentVariables>
+  </PropertyGroup>
   <ItemGroup>
     <Reference Include="System" />
     <Reference Include="System.Xml" />
