+++ Apps/Contoso.Forms.Demo/Contoso.Forms.Demo/App.xaml.cs
@@ -19,12 +19,6 @@ namespace Contoso.Forms.Demo
         void ClearCrashButton();
     }
 
-    public interface IAppConfiguration
-    {
-        string GetAppSecret();
-        string GetTargetToken();
-    }
-
     public partial class App
     {
         public const string LogTag = "AppCenterXamarinDemo";
