+++ Apps/Contoso.Forms.Puppet/Contoso.Forms.Puppet.Droid/Contoso.Forms.Puppet.Droid.csproj
@@ -66,6 +66,9 @@
     <AndroidUseSharedRuntime>false</AndroidUseSharedRuntime>
     <AndroidLinkTool>r8</AndroidLinkTool>
   </PropertyGroup>
+  <PropertyGroup Condition=" '$(RunConfiguration)' == 'Default' ">
+    <Run>am start -e "XAMARIN_FORMS_ANDROID_INT" "app-secret" -e "XAMARIN_FORMS_ANDROID_TARGET_TOKEN_INT" "target-token" -n "com.microsoft.appcenter.xamarin.forms.puppet/Contoso.Forms.Puppet.Droid.MainActivity"</Run>
+  </PropertyGroup>
   <ItemGroup>
     <Reference Include="System" />
     <Reference Include="System.Xml" />
