+++ Apps/Contoso.Forms.Puppet/Contoso.Forms.Puppet/IAppConfiguration.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace Contoso.Forms.Puppet
+{
+    public interface IAppConfiguration
+    {
+        string GetAppSecret();
+        string GetTargetToken();
+    }
+}
